Update README.md
Med.voll.api is a RESTful API developed during a Java and Spring Boot training course provided by Alura, the largest tech platform in Brazil. This project aims to build the backend for a fictional medical clinic. Within this framework, the clinic would utilize the application to manage the registration of doctors and patients as well as the scheduling and cancellation of appointments.